Orbit Home template complete
Finished off the Orbit Home template as well as the footer. Saw some
weird things going on with some of my images, so I've update the style
for the block grid.

Starting on the gallery post type "Clearing" template. Still need to
cleanup the standard templates so they are in the grid correctly.